Commit #5
In this commit, I added the second feature from the rubric which was to visualize data in a user-friendly way, I did this by creating a line-chart in javascript that shows how we price projects at my dad's company. It gives the customer a visual of how the options we provide them will affect the price of their project.